New ClosureData struct
g_cclosure_marshal_VOID__VOID passes two arguments to the closure function: The closure.data (set when the closure is created, called the instance) and a second passed in a GValue. (g_closure_invoke is just a conduit function for the marshaller set on the closure).

Most of the GSignal functions get this wrong, and copy closure.data into the GValue.

We set up a structure with the closure and the pointer and use the pointer to create the GValue.