Change signature of gtk_application_add_app_menu_item to take a label, closure, and data pointer instead of a menu_item.
The App menu doesn't change with the rest of the menubar when the window changes, and its items have to be independent of windows and their menu items. This allows one to associate a callback directly with the app menu item so that it can be invoked directly from the app menu instead of being mediated by a GtkMenuItem.