Add signal handler to override gtk_widget_can_activate_accel for the hidden menubar; makes menu accelerators work.
Gtk by default disables accelerators on GtkMenuItems which aren't visible -- but it provides a signal for overriding this. By providing a signal handler to the menubar (which is what's hidden because we're displaying our own), we enable the accelerators. This wasn't a problem in ige-mac-menu, because it intercepts the NSEvents and dispatches them itself.