Provide for resyncing an already-set menubar.
This is done at two levels: A function gtk_application_sync_menubar is provided for client applications to call from C, but it just asks the currently-active menubar to resync. That's implemented by a new function GNSMenuBar -resync. For it to work, GNSMenubar needs to know what GtkMenuBar it's associated with, so a new private instance variable and a new initializer are provided as well. GNSMenuBar's instance variable app_menu_groups is made private as well, just for cleanliness.