Disconnect the old signal instead of ignoring the fact that the coocoa_item has changed.
It often happens in cases when adding or removing menu items that a new cocoa_menu_item will be created and attached. The original version of this would check to see if there was already an item set on the GtkMenuItem and if there was, would leave the old signal connected and pointing to the wrong NSMenuItem. This change disconnects the old signals and connects new ones with the right data so that the handler can invoke the right Cocoa object.