Rename GtkApplication to GtkOSXApplication
There has been some discussion on gtk-dev about creating one or more application classes, and GtkApplication is one of the proposed names. Since very little of this class has anything to do with anything other than Cocoa at this point, it seems prudent to change the name and avoid a potential collision if a GtkApplication class ever becomes part of Gtk.