Change return type of gtk_osxapplication_get_bundle_info to gchar*
Returning an NSObject* was a dumb idea, because it forces users to use ObjC to deal with it. There are only two likely return values from an Info.plist key anyway, a string (which this now accomodates) and a dictionary (which it doesn't, and which would be difficult to because of the possibility that the dictionary contains more dictionaries).