Split setting up of the Help and Window menus as separate functions.
Adding them without asking was rude and added strings exposed to the user. No provision had been made for translating those strings -- and to do so would expose application ports to an unexpected requirement.

Now, if the application has a Window or Help menu, the application can designate it so that OSX will manage it correctly. If not, and if the application doesn't need translations, the menus can be created -- but this is intended mostly for testing rather than actual use.