Prevent reentry of signal emission in applicationShouldTermintate
GObject signal emissions are not reentrant, so if 
the program is in a handler and the program tries 
to emit another instance of the signal, that emission 
returns with a null result. In our case, that means 
that if one is handling a shouldTerminate and the
user manages to issue another one (perhaps with the
dock menu, which isn't affected by modal dialogs)
the second shouldTerminate will receive an 
NSTerminateNow and the app will quit, perhaps while
trying to save its file. Not so good.

This change prevents that by returning NSTerminateCancel
if the app is waiting on a handler.