Make package work with "make dist"
This took a couple of changes: 
strings/Makefile had to be redone to use automake instead of having a hand-written Makefile.in
Provision had to be made for building all three python binding directories. For this to work, it's necessary to have both gtk+-2 and gtk+-3 installed along with pygobject-2, pygobject-3 and pygtk (and all the other pygtk dependencies), and to pass --enable-python=all to configure.