Bug 703781 - app menu item disappears when made insensitive
Set the GtkMenuItem visible before calling cocoa_menu_item_add_item
in gtkosx_application_insert_app_menu_item instead of calling setHidden
on the resulting NSMenuItem, and
Override GNSMenuItem isEnabled/setEnabled to operate on the action
property instead of the enabled property. It seems that the enabled
property has no effect upon NSMenuItems in the application menu.
Provide a button to toggle the sensitivity of the preferences menu item
in test-integration.c to test the above changes.