Bug 703675 - GtkRecentChooserMenu: GtkRecentChooser doesn't work
gtk_recent_chooser_get_current_uri always returns first item in list
because it doesn't use menu item actions, it uses the menu itself and
gets the active item.

Work around this by emulating the mouse-tracking in GtkMenu with an
NSMenuDelegate responding to willHighlightItem, which tells the
GNSMenuItem to call gtk_menu_set_active to itself.

This requires keeping a reference to the GtkMenuItem on the GNSMenuItem;
we make it a GWeakRef to avoid a reference loop.

Also release submenus when they're replaced to avoid a leak.