Provide gtk-mac-integration.pc for backwards compatibility.
gtk-mac-integration.pc is a copy of gtk-mac-integration-gtk2.pc unless
only gtk3 is present or configured in which case it's a copy of
gtk-mac-integration-gtk3.pc.

The relaxes the need for applications to fix their configure.ac to update
to gtk-mac-integration-gtk2.pc or gtk-mac-integration-gtk3.pc in
configure.ac: It will work in all cases for gtk2 applications as long as
gtk2 is available and configured when gtk-mac-integration is built, but
will work for gtk3 applications only in those cases where only gtk3 is
present or configured.

It's still advisable for all applications to specify which
gtk-mac-integration they need for cases like MacPorts where they might
be built with both present and with gtk-mac-integration-2.0.7.