Remove the Carbon-based gtk-mac-menu/gtk-mac-dock/gtk-mac-bundle.
These are 32-bit only and 32-bit support was removed from the
macOS 10.15 SDK. The compilation guard specified x86_64 so it
failed for Apple Silicon and tried to build. At this point it makes
more sense to remove the code instead of updating the guard.

Fixes https://gitlab.gnome.org/GNOME/gtk-mac-integration/-/issues/7