Remove PyGtk support.
PyGtk was the preferred python interface for Gtk2, but development
ceased several years ago and it depends on Python2, also no longer
maintainted. A check of MacPorts shows no packages currently depend on
PyGtk bindings for gtk-mac-integration. Gtk2 had its last release in
December 2020. There are no differences in GtkosxApplication required
for C support of Gtk2 and Gtk3 so there's no reason to disable support
for Gtk2 in C, but it's no longer reasonable to maintain the PyGtk
support code.